+++ RogueLike/Src/Random.cpp
@@ -10,6 +10,6 @@ int getRandom(int& seed, int min, int max)
 
 	seed = (std::abs(seed) * 38261 + 58243) % 93871;
 	float procent = std::abs(seed) / 93871.0f;
-	int v = min + (max - min) * (procent);
+	int v = (int)(min + (max - min) * (procent));
 	return std::clamp(v, min, max - 1);
 }
